requirements.txt
slack_bolt 
python-dotenv 
google-generativeai